+++ packages/eth-common/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v2.1.0
+
+- Add toolbar product links from `@alethio/explorer-plugin-eth-extended`. See [config.sample.json](config.sample.json) for configuration changes.
+
 # v2.0.0
 
 - Update search API interface. Search widget now requires eth-lite or eth-extended v2 plugins.